+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
   - `--redis-dsn` redis server DSN (format: `redis://<user>:<password>@<host>:<port>/<db_number>`), required when storage driver `redis` is used
   - `--broadcast-driver` broadcast driver
   - `--storage-driver` storage driver (`redis` and `memory` is supported)
+  - `--max-request-body-size` maximal webhook request body size (in bytes)
 - Sub-command `healthcheck` (hidden in CLI help) that makes a simple HTTP request (with user-agent `HealthChecker/internal`) to the `http://127.0.0.1:8080/live` endpoint. Port number can be changed using `--port`, `-p` flag or `LISTEN_PORT` environment variable
 - Healthcheck in dockerfile
 - Global (available for all sub-commands) flags:
@@ -24,6 +25,7 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 - HTTP requests & HTTP panics logging middlewares
 - Logging using `uber-go/zap` package
 - Webhook delay uses `time.After()` instead `time.Sleep()` with context canceling support
+- HTTP route `/api/version`
 
 ### Changed
 
@@ -40,6 +42,7 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 - HTTP method `CONNECT` for webhook endpoint
 - `serve` sub-command flags:`--redis-host`, `--redis-port`, `--redis-db-num`
 - Environment variables support: `REDIS_HOST`, `REDIS_PORT`, `REDIS_DB_NUM`
+- Property `version` from `/api/settings` JSON-object response
 
 ### Fixed